@@ -1,59 +1,136 @@
-import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { selectDevModeEnabled, selectMockHospitalData } from './devModeSlice';
+import { createSlice, PayloadAction } from "@reduxjs/toolkit";
+import { RootState } from "../store";
 
-interface Doctor {
+export interface Doctor {
   id: string;
   name: string;
   specialization: string;
   licenseNumber?: string;
 }
 
-interface Hospital {
+export interface Hospital {
+  id?: string | number;
   name: string;
   did: string;
-  address?: string;
+  address: string;
   phone?: string;
+  phone_number?: string;
   email?: string;
-  doctors?: Doctor[];
+  account?: string;
+  license_number?: string;
+  doctor_name?: string;
+  authoriazation_number?: string;
+  x_axis?: number;
+  y_axis?: number;
+  public_key?: string | null;
+  doctors: Doctor[];
 }
 
-interface AuthState {
-  isAuthenticated: boolean;
-  user: any;
+export interface User {
+  id: string;
+  account: string;
+  hospitalName: string;
+  hospital?: Hospital;
+}
+
+export interface Tokens {
+  access_token: string;
+  refresh_token: string;
+}
+
+export interface AuthState {
+  user: User | null;
   token: string | null;
+  isLoggedIn: boolean;
   loading: boolean;
   error: string | null;
-  hospital: any;
-  currentDoctor: any;
+  hospital: Hospital | null;
 }
 
 const initialState: AuthState = {
-  isAuthenticated: localStorage.getItem('token') ? true : false,
-  user: JSON.parse(localStorage.getItem('user') || 'null'),
-  token: localStorage.getItem('token'),
+  user: null,
+  token: null,
+  isLoggedIn: false,
   loading: false,
   error: null,
   hospital: null,
-  currentDoctor: null
 };
 
-const authSlice = createSlice({
-  name: 'auth',
-  initialState,
+// 로컬 스토리지에서 기존 데이터 로드
+const loadAuthStateFromStorage = (): AuthState => {
+  try {
+    const token = localStorage.getItem("access_token");
+    const userStr = localStorage.getItem("user");
+    const hospitalStr = localStorage.getItem("hospital");
+
+    let user = null;
+    let hospital = null;
+
+    if (userStr) {
+      user = JSON.parse(userStr);
+    }
+
+    if (hospitalStr) {
+      hospital = JSON.parse(hospitalStr);
+    }
+
+    if (token) {
+      return {
+        user,
+        token,
+        hospital,
+        isLoggedIn: true,
+        loading: false,
+        error: null,
+      };
+    }
+  } catch (e) {
+    console.error("로컬 스토리지에서 인증 상태 로드 중 오류:", e);
+  }
+
+  return initialState;
+};
+
+const preloadedState = loadAuthStateFromStorage();
+
+export const authSlice = createSlice({
+  name: "auth",
+  initialState: preloadedState,
   reducers: {
-    signupStart: (state) => {
-      state.loading = true;
-      state.error = null;
+    setCredentials: (
+      state,
+      action: PayloadAction<{ hospital: Hospital; tokens: Tokens }>
+    ) => {
+      state.hospital = action.payload.hospital;
+
+      //id, 계정, 병원 이름
+      state.user = {
+        id: action.payload.hospital.id
+          ? String(action.payload.hospital.id)
+          : "",
+        account: action.payload.hospital.account || "",
+        hospitalName: action.payload.hospital.name,
+        hospital: action.payload.hospital,
+      };
+      state.token = action.payload.tokens.access_token;
+      state.isLoggedIn = true;
+
+      // 로컬 스토리지 업데이트 에세스토큰 물어보기기
+      localStorage.setItem("access_token", action.payload.tokens.access_token);
+      localStorage.setItem("hospital", JSON.stringify(action.payload.hospital));
+      localStorage.setItem("user", JSON.stringify(state.user));
     },
-    signupSuccess: (state, action: PayloadAction<any>) => {
-      state.loading = false;
-      state.error = null;
-      state.user = action.payload;
-      state.isAuthenticated = true;
-    },
-    signupFailure: (state, action: PayloadAction<string>) => {
-      state.loading = false;
-      state.error = action.payload;
+    logout: (state) => {
+      state.user = null;
+      state.token = null;
+      state.isLoggedIn = false;
+      state.hospital = null;
+
+      // 로컬 스토리지에서 인증 데이터 제거
+      localStorage.removeItem("access_token");
+      localStorage.removeItem("refresh_token");
+      localStorage.removeItem("user");
+      localStorage.removeItem("hospital");
     },
     registerStart: (state) => {
       state.loading = true;
@@ -62,80 +139,58 @@ const authSlice = createSlice({
     registerSuccess: (state, action: PayloadAction<any>) => {
       state.loading = false;
       state.error = null;
-      state.user = action.payload;
-      state.isAuthenticated = true;
     },
     registerFailure: (state, action: PayloadAction<string>) => {
       state.loading = false;
       state.error = action.payload;
     },
-    loginStart: (state) => {
-      state.loading = true;
+    clearError: (state) => {
       state.error = null;
     },
-    loginSuccess: (state, action: PayloadAction<{user: any, token: string, hospital?: any}>) => {
-      state.isAuthenticated = true;
-      state.user = action.payload.user;
-      state.token = action.payload.token;
-      state.loading = false;
-      state.error = null;
-      state.hospital = action.payload.hospital || null;
-      
-      localStorage.setItem('user', JSON.stringify(action.payload.user));
-      localStorage.setItem('token', action.payload.token);
-    },
-    loginFailure: (state, action: PayloadAction<string>) => {
-      state.loading = false;
-      state.error = action.payload;
-    },
-    logout: (state) => {
-      state.isAuthenticated = false;
-      state.user = null;
-      state.token = null;
-      state.hospital = null;
-      state.currentDoctor = null;
-      
-      localStorage.removeItem('user');
-      localStorage.removeItem('token');
-    },
-    updateHospitalInfo: (state, action: PayloadAction<any>) => {
-      state.hospital = { ...state.hospital, ...action.payload };
-      if (state.user && state.hospital && state.user.id === state.hospital.id) {
-        state.user = { ...state.user, ...action.payload };
-        localStorage.setItem('user', JSON.stringify(state.user));
+    setHospitalInfo: (state, action: PayloadAction<Hospital>) => {
+      state.hospital = action.payload;
+
+      // 로컬 스토리지에 병원 정보 업데이트
+      localStorage.setItem("hospital", JSON.stringify(action.payload));
+
+      // 사용자 데이터에 병원 이름 업데이트
+      if (state.user) {
+        state.user.hospitalName = action.payload.name;
+        localStorage.setItem("user", JSON.stringify(state.user));
       }
     },
-    setCurrentDoctor: (state, action: PayloadAction<any>) => {
-      state.currentDoctor = action.payload;
+    updateHospitalInfoAction: (state, action: PayloadAction<Hospital>) => {
+      state.hospital = action.payload;
+
+      // 로컬 스토리지에 병원 정보 업데이트
+      localStorage.setItem("hospital", JSON.stringify(action.payload));
+
+      // 사용자 데이터에 병원 이름 업데이트
+      if (state.user) {
+        state.user.hospitalName = action.payload.name;
+        localStorage.setItem("user", JSON.stringify(state.user));
+      }
     },
   },
 });
 
 export const {
-  signupStart,
-  signupSuccess,
-  signupFailure,
+  setCredentials,
+  logout,
   registerStart,
   registerSuccess,
   registerFailure,
-  loginStart,
-  loginSuccess,
-  loginFailure,
-  logout,
-  updateHospitalInfo,
-  setCurrentDoctor
+  clearError,
+  updateHospitalInfoAction,
+  setHospitalInfo,
 } = authSlice.actions;
 
-export const selectIsAuthenticated = (state: { auth: AuthState }) => state.auth.isAuthenticated;
-export const selectUser = (state: { auth: AuthState }) => state.auth.user;
-export const selectToken = (state: { auth: AuthState }) => state.auth.token;
-export const selectLoggedInHospital = (state: any) => {
-  const devModeEnabled = selectDevModeEnabled(state);
-  if (devModeEnabled && state.auth.user?.email === 'ssafy@example.com') {
-    return selectMockHospitalData(state);
-  }
-  return state.auth.hospital;
-};
-export const selectCurrentDoctor = (state: { auth: AuthState }) => state.auth.currentDoctor;
+// 선택자(Selectors)
+export const selectCurrentUser = (state: RootState) => state.auth.user;
+export const selectCurrentToken = (state: RootState) => state.auth.token;
+export const selectIsLoggedIn = (state: RootState) => state.auth.isLoggedIn;
+export const selectLoading = (state: RootState) => state.auth.loading;
+export const selectError = (state: RootState) => state.auth.error;
+export const selectLoggedInHospital = (state: RootState) => state.auth.hospital;
 
-export default authSlice.reducer; 
+export default authSlice.reducer;
